+++ malleefowl/esgf/search.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from dateutil import parser as date_parser
 
 import threading
 from Queue import Queue
@@ -137,21 +136,25 @@ class ESGSearch(object):
         # TODO: check type of start, end
         logger.debug('start=%s, end=%s', start, end)
 
-        start_date = end_date = None
-        # TODO: maybe only use start/end if temporal=True
-        if start is not None and end is not None:
-            start_date = date_parser.parse(start)
-            end_date = date_parser.parse(end)
-
         ctx = None
         if temporal is True:
             logger.debug("using dataset search with time constraints")
+            # TODO: handle timestamps in a better way
+            timestamp_format = '%Y-%m-%dT%H:%M:%SZ'
+            if start:
+                from_timestamp = start.strftime(timestamp_format)
+            else:
+                from_timestamp = None
+            if end:
+                to_timestamp = end.strftime(timestamp_format)
+            else:
+                to_timestamp = None
             ctx = self.conn.new_context(fields=self.fields,
                                         replica=self.replica,
                                         latest=self.latest,
                                         query=query,
-                                        from_timestamp=start,
-                                        to_timestamp=end)
+                                        from_timestamp=from_timestamp,
+                                        to_timestamp=to_timestamp)
         else:
             ctx = self.conn.new_context(fields=self.fields,
                                         replica=self.replica,
@@ -176,7 +179,7 @@ class ESGSearch(object):
         self.count = 0
         # search datasets
         # we always do this to get the summary document
-        datasets = ctx.search()
+        datasets = ctx.search(ignore_facet_check=True)
 
         (self.start_index, self.stop_index, self.max_count) = self._index(datasets, limit, offset)
         self.summary['number_of_datasets'] = max(0, self.max_count)
@@ -201,7 +204,7 @@ class ESGSearch(object):
             pass
         # search files (optional)
         elif search_type == 'File':
-            self._file_search(datasets, my_constraints, start_date, end_date)
+            self._file_search(datasets, my_constraints, start, end)
         # search aggregations (optional)
         elif search_type == 'Aggregation':
             self._aggregation_search(datasets, my_constraints)
@@ -228,7 +231,7 @@ class ESGSearch(object):
             ctx = self.local_ctx.constrain(instance_id=dataset.json.get('instance_id'))
             if ctx.hit_count == 1:
                 logger.info('found local replica')
-                datasets = ctx.search()
+                datasets = ctx.search(ignore_facet_check=True)
                 f_ctx = datasets[0].file_context()
             else:
                 logger.info('no local replica found')
@@ -242,7 +245,7 @@ class ESGSearch(object):
             ctx = self.local_ctx.constrain(instance_id=dataset.json.get('instance_id'))
             if ctx.hit_count == 1:
                 logger.info('found local replica')
-                datasets = ctx.search()
+                datasets = ctx.search(ignore_facet_check=True)
                 agg_ctx = datasets[0].aggregation_context()
             else:
                 logger.info('no local replica found')
@@ -265,7 +268,7 @@ class ESGSearch(object):
     def _file_search_job(self, f_ctx, start_date, end_date):
         # logger.debug('num files: %d', f_ctx.hit_count)
         logger.debug('facet constraints=%s', f_ctx.facet_constraints)
-        for f in f_ctx.search():
+        for f in f_ctx.search(ignore_facet_check=True):
             if not temporal_filter(f.filename, start_date, end_date):
                 continue
             with self.result_lock:
@@ -340,7 +343,7 @@ class ESGSearch(object):
             if agg_ctx.hit_count == 0:
                 logger.warn('dataset %s has no aggregations!', ds.dataset_id)
                 continue
-            for agg in agg_ctx.search():
+            for agg in agg_ctx.search(ignore_facet_check=True):
                 self.summary['number_of_selected_aggregations'] = self.summary['number_of_selected_aggregations'] + 1
                 self.summary['aggregation_size'] = self.summary['aggregation_size'] + agg.size
                 self.result.append(agg.opendap_url)