+++ malleefowl/processes/wps_download.py
@@ -1,79 +1,87 @@
 import json
-from pywps.Process import WPSProcess
+
+from pywps import Process
+from pywps import LiteralInput
+from pywps import ComplexInput
+from pywps import LiteralOutput
+from pywps import ComplexOutput
+from pywps import Format, FORMATS
+from pywps.app.Common import Metadata
 
 from malleefowl.download import download_files
 
+import logging
+LOGGER = logging.getLogger("PYWPS")
 
-class Download(WPSProcess):
+
+class Download(Process):
 
     def __init__(self):
-        WPSProcess.__init__(
-            self,
+        inputs = [
+            LiteralInput('resource', 'Resource',
+                         data_type='string',
+                         abstract="URL of your resource.",
+                         min_occurs=1,
+                         max_occurs=1024,
+                         ),
+            ComplexInput('credentials', 'X509 Certificate',
+                         abstract='Optional X509 proxy certificate to access ESGF data.'
+                         'This parameter is deprecated. Use X-X509-User-Proxy header variable.',
+                         metadata=[Metadata('Info')],
+                         min_occurs=0,
+                         max_occurs=1,
+                         supported_formats=[Format('application/x-pkcs7-mime')]),
+        ]
+        outputs = [
+            ComplexOutput('output', 'Downloaded files',
+                          abstract="Json document with list of downloaded files with file url.",
+                          as_reference=True,
+                          supported_formats=[Format('application/json')]),
+        ]
+
+        super(Download, self).__init__(
+            self._handler,
             identifier="download",
             title="Download files",
-            version="0.6",
+            version="0.7",
             abstract="Downloads files and provides file list as json document.",
-            statusSupported=True,
-            storeSupported=True)
-
-        self.resource = self.addLiteralInput(
-            identifier="resource",
-            title="Resource",
-            abstract="URL of your resource ...",
-            minOccurs=1,
-            maxOccurs=2048,
-            type=type('')
+            metadata=[
+                Metadata('Birdhouse', 'http://bird-house.github.io/'),
+                Metadata('User Guide', 'http://malleefowl.readthedocs.io/en/latest/'),
+            ],
+            inputs=inputs,
+            outputs=outputs,
+            status_supported=True,
+            store_supported=True,
         )
 
-        self.credentials = self.addComplexInput(
-            identifier="credentials",
-            title="X509 Certificate",
-            abstract="X509 proxy certificate to access ESGF data.",
-            minOccurs=0,
-            maxOccurs=1,
-            maxmegabites=1,
-            formats=[{"mimeType": "application/x-pkcs7-mime"}],
-        )
+    def _handler(self, request, response):
+        response.update_status("starting download ...", 0)
 
-        self.openid = self.addLiteralInput(
-            identifier="openid",
-            title="ESGF OpenID",
-            abstract="For example: https://esgf-data.dkrz.de/esgf-idp/openid/username",
-            minOccurs=0,
-            maxOccurs=1,
-            type=type('')
-        )
+        urls = [resource.data for resource in request.inputs['resource']]
+        LOGGER.debug("downloading urls: %s", len(urls))
 
-        self.password = self.addLiteralInput(
-            identifier="password",
-            title="OpenID Password",
-            abstract="Enter your Password",
-            minOccurs=0,
-            maxOccurs=1,
-            type=type('')
-        )
+        if 'X-X509-User-Proxy' in request.http_request.headers:
+            credentials = request.http_request.headers['X-X509-User-Proxy']
+            LOGGER.debug('Using X509_USER_PROXY.')
+        elif 'credentials' in request.inputs:
+            credentials = request.inputs['credentials'][0].file
+            LOGGER.warn('Using deprecated input parameter credentials.')
+        else:
+            credentials = None
+            LOGGER.debug('Using no credentials')
 
-        self.output = self.addComplexOutput(
-            identifier="output",
-            title="Downloaded files",
-            abstract="Json document with list of downloaded files with file url.",
-            metadata=[],
-            formats=[{"mimeType": "application/json"}],
-            asReference=True,
-        )
+        def monitor(msg, progress):
+            response.update_status(msg, progress)
 
-    def monitor(self, msg, progress):
-        self.status.set(msg, progress)
-
-    def execute(self):
         files = download_files(
-            urls=self.getInputValues(identifier='resource'),
-            credentials=self.credentials.getValue(),
-            openid=self.openid.getValue(),
-            password=self.password.getValue(),
-            monitor=self.monitor)
+            urls=urls,
+            credentials=credentials,
+            monitor=monitor)
 
-        outfile = 'out.json'
-        with open(outfile, 'w') as fp:
+        with open('out.json', 'w') as fp:
             json.dump(obj=files, fp=fp, indent=4, sort_keys=True)
-        self.output.setValue(outfile)
+            response.outputs['output'].file = fp.name
+
+        response.update_status("download done", 100)
+        return response